Revert "수정페이지 추가 관련 PR입니다.(#11)"
This reverts commit 3d71c2d18f97d67ca9226b15ed5ad1f446443391.
