+++ src/pages/MyPage/MyPage.jsx
@@ -1,10 +1,9 @@
-import MyPageHeader from '../../components/Header/MyPage';
-import React from 'react';
-import Profile from '../../components/MyPage/Profile';
-import Modify from '../../components/MyPage/Modify';
 import styled from 'styled-components';
-import { PostingList } from '../../components/MyPage/PostingList';
+import MyPageHeader from '../../components/Header/MyPage';
 import { Calendar } from '../../components/MyPage/Calendar';
+import Modify from '../../components/MyPage/Modify';
+import { PostingList } from '../../components/MyPage/PostingList';
+import Profile from '../../components/MyPage/Profile';
 
 export default function MyPage() {
   return (
@@ -14,12 +13,12 @@ export default function MyPage() {
         <StMypageTopWrap>
           <LeftColumn>
             <Profile />
-          </LeftColumn>
-          <RightColumn>
-            {/* <Modify /> */}
             <StCalendarSection>
               <Calendar />
             </StCalendarSection>
+          </LeftColumn>
+          <RightColumn>
+            <Modify />
           </RightColumn>
         </StMypageTopWrap>
         <StPostingListTitle>Posting List</StPostingListTitle>
@@ -34,20 +33,19 @@ export default function MyPage() {
 const LeftColumn = styled.div`
   display: flex;
   flex-direction: column;
+  align-items: center;
   justify-content: center;
-  gap: 30px;
 `;
 
 const RightColumn = styled.div`
   display: flex;
   flex-direction: column;
-  align-items: center;
   justify-content: center;
 `;
 
 const StCalendarSection = styled.div`
-  width: 700px;
-  height: 200px;
+  width: 400px;
+  height: 150px;
   cursor: pointer;
 `;
 
@@ -59,7 +57,7 @@ const StMypageWrapper = styled.div`
 
 const StMypageTopWrap = styled.div`
   display: flex;
-  gap: 6rem;
+  gap: 2rem;
   align-items: center;
   justify-content: center;
   padding: 4rem;
