+++ src/pages/WritePage/WritePage.jsx
@@ -1,5 +1,5 @@
 import { useAuth } from '../../contexts/Auth/hooks';
-import WriteHeader from '../../components/Header/Modify';
+import WriteHeader from '../../components/Header/Write';
 
 export default function WritePage() {
   const { supabaseClient, session } = useAuth();
@@ -16,19 +16,19 @@ export default function WritePage() {
 
   return (
     <>
-      <WriteHeader />
-      <div>
-        <form>
-          <button
-            onClick={(e) => {
-              e.preventDefault();
-              handlePosting();
-            }}
-          >
-            글 쓰기
-          </button>
-        </form>
-      </div>
+    <WriteHeader />
+    <div>
+      <form>
+        <button
+          onClick={(e) => {
+            e.preventDefault();
+            handlePosting();
+          }}
+        >
+          글 쓰기
+        </button>
+      </form>
+    </div>
     </>
   );
 }
