+++ src/routes/Router.jsx
@@ -3,7 +3,6 @@ import HomePage from '../pages/HomePage/HomePage';
 import LoginPage from '../pages/LoginPage/LoginPage';
 import MyPage from '../pages/MyPage/MyPage';
 import WritePage from '../pages/WritePage/WritePage';
-import ModifyPage from '../pages/ModifyPage/ModifyPage';
 
 export default function Router() {
   return (
@@ -11,7 +10,6 @@ export default function Router() {
       <Routes>
         <Route path="/" element={<HomePage />} />
         <Route path="/write" element={<WritePage />} />
-        <Route path="/ModifyPage" element={<ModifyPage />} />
         <Route path="/:userName/posts" element={<MyPage />} />
         <Route path="/login" element={<LoginPage />} />
       </Routes>
